restructured the codebase for better scaling
