+++ frontend_only/frontend_scripts/next_frontendNoAuth.go
@@ -1,4 +1,4 @@
-package next_boil
+package frontend_only_scripts
 
 import (
 	"fmt"
@@ -8,7 +8,7 @@ import (
 	"sam.crider/boilerplate-script/utils"
 )
 
-func Next_FrontendClerk(project_name string) {
+func Next_FrontendNoAuth(project_name string) {
 
 	// create next app
 	cmd := utils.BoundCommand("npx", "create-next-app", project_name, "--typescript")
@@ -91,25 +91,14 @@ func Next_FrontendClerk(project_name string) {
 		}
 	}
 
-	// remove readme and replace with clerk readme
+	// remove readme and replace with no auth readme
 	err = os.Remove("README.md")
 	if err != nil {
 		fmt.Println(err)
 		return
 	}
 
-	utils.Create_File("README.md", generated.File__nextClerkFrontReadMe)
-
-	// install deps
-	cmd_deps := utils.BoundCommand("npm", "install", "@clerk/nextjs")
-
-	if err := cmd_deps.Run(); err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	// make .env.local file
-	utils.Create_File(".env.local", generated.File__nextClerkEnvLocal)
+	utils.Create_File("README.md", generated.File__nextNoAuthFrontReadMe)
 
 	// replace the gitignore file
 	err = os.Remove(".gitignore")
@@ -123,32 +112,6 @@ func Next_FrontendClerk(project_name string) {
 	utils.Work_wrapper(func() {
 		// cd src directory
 		err = os.Chdir("src")
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		// create clerk middleware file
-		utils.Create_File("middleware.ts", generated.File__nextClerkMiddleware)
-
-		// cd app directory
-		err = os.Chdir("app")
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		// remove and replace the layout file
-		err = os.Remove("layout.tsx")
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		utils.Create_File("layout.tsx", generated.File__nextClerkLayout)
-
-		// cd out of app
-		err = os.Chdir("..")
 		if err != nil {
 			fmt.Println(err)
 			return
