+++ full_stack/next/next_firebase.go
@@ -1,4 +1,4 @@
-package next_boil
+package next_scripts
 
 import (
 	"fmt"
